+++ tests/tests.py
@@ -155,3 +155,173 @@ class AuthTestCase(test.APITestCase):
         self.assertFalse(
             AuthToken.objects.exists(),
             'Encrypted tokens were not erased on reset')
+
+    def test_tokens_endpoint(self):
+        """
+        The tokens endpoint supports general token management.
+        """
+        creds = dict(login=dict(
+            username='foo-username', password='foo-secret'))
+        invalid_creds = dict(login=dict(
+            username='foo-username', password='bar-secret'))
+        user = User.objects.create_user(**creds['login'])
+        self.client.credentials(HTTP_AUTHORIZATION=get_basic_auth_header(
+            creds['login']['username'], creds['login']['password']))
+        token = AuthToken.objects.create(
+            user=user, password=creds['login']['password'])
+        tokens = AuthToken.objects.filter(user=user).order_by('created')
+        token_obj = tokens.get()
+        token_json = dict(
+            token=token,
+            created=token_obj.created.replace(tzinfo=None).isoformat() + 'Z',
+            expires=token_obj.expires.replace(tzinfo=None).isoformat() + 'Z')
+
+        # HTTP methods that may ignore entity bodies are not allowed,
+        # with or without credentials included
+        for method in ('get', 'delete'):
+            response = getattr(self.client, method)(
+                '/api/auth/tokens/', format='json')
+            self.assertEqual(
+                response.status_code, 405,
+                'GET method should not be supported')
+            response = getattr(self.client, method)(
+                '/api/auth/tokens/{0}/'.format(token_obj.pk), format='json')
+            self.assertEqual(
+                response.status_code, 405,
+                'GET method should not be supported')
+            response = getattr(self.client, method)(
+                '/api/auth/tokens/', creds, format='json')
+            self.assertEqual(
+                response.status_code, 405,
+                'GET method should not be supported')
+            response = getattr(self.client, method)(
+                '/api/auth/tokens/{0}/'.format(token_obj.pk),
+                creds, format='json')
+            self.assertEqual(
+                response.status_code, 405,
+                'GET method should not be supported')
+
+        # Only POST is allowed for the endpoints that replace those methods,
+        # with or without credentials included
+        for method in ('put', 'patch'):
+            response = getattr(self.client, method)(
+                '/api/auth/tokens/list/', format='json')
+            self.assertEqual(
+                response.status_code, 405,
+                '{0} method should not be supported'.format(method.upper()))
+            response = getattr(self.client, method)(
+                '/api/auth/tokens/list/', creds, format='json')
+            self.assertEqual(
+                response.status_code, 405,
+                '{0} method should not be supported'.format(method.upper()))
+            for detail_action in ('retrieve', 'destroy'):
+                response = getattr(self.client, method)(
+                    '/api/auth/tokens/{0}/{1}/'.format(
+                        token_obj.pk, detail_action), format='json')
+                self.assertEqual(
+                    response.status_code, 405,
+                    '{0} method should not be supported'.format(
+                        method.upper()))
+                response = getattr(self.client, method)(
+                    '/api/auth/tokens/{0}/{1}/'.format(
+                        token_obj.pk, detail_action), creds, format='json')
+                self.assertEqual(
+                    response.status_code, 405,
+                    '{0} method should not be supported'.format(
+                        method.upper()))
+
+        # Allowed endpoints and methods all require credentials
+        response = self.client.post('/api/auth/tokens/', format='json')
+        self.assertEqual(
+            response.status_code, 400,
+            'Wrong response to missing credentials')
+        response = self.client.post(
+            '/api/auth/tokens/', invalid_creds, format='json')
+        self.assertEqual(
+            response.status_code, 400,
+            'Wrong response to invalid credentials')
+        for method in ('put', 'patch'):
+            response = getattr(self.client, method)(
+                '/api/auth/tokens/{0}/'.format(token_obj.pk),
+                format='json')
+            self.assertEqual(
+                response.status_code, 400,
+                'Wrong response to missing credentials')
+            response = getattr(self.client, method)(
+                '/api/auth/tokens/{0}/'.format(token_obj.pk),
+                invalid_creds, format='json')
+            self.assertEqual(
+                response.status_code, 400,
+                'Wrong response to invalid credentials')
+        response = self.client.post(
+            '/api/auth/tokens/list/', format='json')
+        self.assertEqual(
+            response.status_code, 400,
+            'Wrong response to missing credentials')
+        response = self.client.post(
+            '/api/auth/tokens/list/', invalid_creds, format='json')
+        self.assertEqual(
+            response.status_code, 400,
+            'Wrong response to invalid credentials')
+        for detail_action in ('retrieve', 'destroy'):
+            response = self.client.post(
+                '/api/auth/tokens/{0}/{1}/'.format(
+                    token_obj.pk, detail_action),
+                format='json')
+            self.assertEqual(
+                response.status_code, 400,
+                'Wrong response to missing credentials')
+            response = self.client.post(
+                '/api/auth/tokens/{0}/{1}/'.format(
+                    token_obj.pk, detail_action),
+                invalid_creds, format='json')
+            self.assertEqual(
+                response.status_code, 400,
+                'Wrong response to invalid credentials')
+
+        list_response = self.client.post(
+            '/api/auth/tokens/list/', creds, format='json')
+        self.assertEqual(
+            list_response.status_code, 200, 'List request did not succeed')
+        self.assertIsInstance(
+            list_response.json(), list, 'Wrong type returned')
+        self.assertEqual(
+            len(list_response.json()), 1, 'Wrong number of results')
+        self.assertEqual(
+            list_response.json()[0], token_json, 'Wrong token JSON')
+
+        retrieve_response = self.client.post(
+            '/api/auth/tokens/{0}/retrieve/'.format(token_obj.pk),
+            creds, format='json')
+        self.assertEqual(
+            retrieve_response.status_code, 200,
+            'Retrieve request did not succeed')
+        self.assertIsInstance(
+            retrieve_response.json(), dict, 'Wrong type returned')
+        self.assertEqual(
+            retrieve_response.json(), token_json, 'Wrong token JSON')
+
+        create_response = self.client.post(
+            '/api/auth/tokens/', creds, format='json')
+        self.assertEqual(
+            create_response.status_code, 201,
+            'Create request did not succeed')
+        self.assertIsInstance(
+            create_response.json(), dict, 'Wrong type returned')
+        self.assertEqual(tokens.count(), 2, 'Wrong number of tokens')
+        self.assertEqual(
+            create_response.json(), dict(
+                token=tokens[1].decrypt(creds['login']['password']),
+                created=tokens[1].created.replace(
+                    tzinfo=None).isoformat() + 'Z',
+                expires=tokens[1].expires.replace(
+                    tzinfo=None).isoformat() + 'Z'),
+            'Wrong created token JSON')
+
+        destroy_response = self.client.post(
+            '/api/auth/tokens/{0}/destroy/'.format(tokens[1].pk),
+            creds, format='json')
+        self.assertEqual(
+            destroy_response.status_code, 204,
+            'Destroy request did not succeed')
+        self.assertEqual(tokens.count(), 1, 'Wrong number of tokens')